refactor(core): change Button component to typescript
